pyproject.toml
[tool.poetry]
name = "gamebattle-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "gamebattle_backend"}]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
mypy = "^1.5.1"
pyyaml = "^6.0.1"
pydantic = "^2.3.0"
fastapi = "^0.103.1"
firebase-admin = "^6.2.0"
websockets = "^11.0.3"
trio = "^0.22.2"
httpx = "^0.25.0"
redis = "^5.0.1"
groq = "^0.11.0"
aiodocker = "^0.24.0"


[tool.poetry.group.dev.dependencies]
types-pyyaml = "^6.0.12.20240917"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
